backend/requirements.txt
Flask
requests
python-dotenv
sumeval
spacy
sacrebleu==1.5.1
BeautifulSoup4
